requirements.txt
python-telegram-bot==21.6
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.126.0
gspread==6.1.2
google-generativeai==0.7.2
python-dotenv==1.0.1
aiohttp>=3.9.0
aiofiles>=23.2.0
Pillow>=10.0.0